add locks to annotate command

+++ src/commands/22_annotate/raven-cmd-annotate.adb
@@ -10,6 +10,7 @@ with Raven.Database.Operations;
 with Raven.Database.Annotate;
 with Raven.Database.Search;
 with Raven.Database.Query;
+with Raven.Database.Lock;
 
 use Raven.Strings;
 
@@ -19,6 +20,7 @@ package body Raven.Cmd.Annotate is
    package QRY renames Raven.Database.Query;
    package OPS renames Raven.Database.Operations;
    package ANN renames Raven.Database.Annotate;
+   package LOK renames Raven.Database.Lock;
    package RCU renames Raven.Cmd.Unset;
 
 
@@ -30,7 +32,38 @@ package body Raven.Cmd.Annotate is
       rdb : Database.RDB_Connection;
       unfinished_packages : Pkgtypes.Package_Set.Vector;
       behave_cs : Boolean := comline.common_options.case_sensitive;
+      active_lock : LOK.lock_type := LOK.lock_readonly;
 
+      procedure release_active_lock is
+      begin
+         if not LOK.release_lock (rdb, active_lock) then
+            case active_lock is
+               when LOK.lock_advisory  => Event.emit_error (LOK.no_advisory_unlock);
+               when LOK.lock_exclusive => Event.emit_error (LOK.no_exclusive_unlock);
+               when LOK.lock_readonly  => Event.emit_error (LOK.no_read_unlock);
+            end case;
+         end if;
+      end release_active_lock;
+
+      procedure exit_lock is
+      begin
+         release_active_lock;
+         OPS.rdb_close (rdb);
+      end exit_lock;
+
+      function activate_lock return Boolean is
+      begin
+         if not LOK.obtain_lock (rdb, active_lock) then
+            case active_lock is
+               when LOK.lock_advisory  => Event.emit_error (LOK.no_adv_lock);
+               when LOK.lock_exclusive => Event.emit_error (LOK.no_exc_lock);
+               when LOK.lock_readonly  => Event.emit_error (LOK.no_read_lock);
+            end case;
+            OPS.rdb_close (rdb);
+            return False;
+         end if;
+         return True;
+      end activate_lock;
    begin
       if Context.reveal_case_sensitive then
          behave_cs := True;
@@ -40,6 +73,10 @@ package body Raven.Cmd.Annotate is
          when RESULT_OK => null;
          when others => return False;
       end case;
+
+      if not activate_lock then
+         return False;
+      end if;
 
       Database.Search.rvn_core_search
         (db           => rdb,
@@ -54,15 +91,34 @@ package body Raven.Cmd.Annotate is
 
       if comline.cmd_annotate.operation_find then
          display_tags (rdb, unfinished_packages, USS (comline.cmd_annotate.tag));
+         exit_lock;
+
       elsif comline.cmd_annotate.operation_delete then
+
+         exit_lock;
+         active_lock := LOK.lock_advisory;
+         if not activate_lock then
+            return False;
+         end if;
+
          delete_tags (rdb, unfinished_packages, USS (comline.cmd_annotate.tag),
                       comline.common_options.quiet);
+         exit_lock;
+
       elsif comline.cmd_annotate.operation_set then
+
+         exit_lock;
+         active_lock := LOK.lock_advisory;
+         if not activate_lock then
+            return False;
+         end if;
+
          define_tags (rdb, unfinished_packages, USS (comline.cmd_annotate.tag),
                       USS (comline.cmd_annotate.note), comline.common_options.quiet);
+         exit_lock;
+
       end if;
 
-      OPS.rdb_close (rdb);
       return True;
    end execute_annotate_command;
